+++ app/src/androidTest/java/br/com/lucas/todo/ExampleInstrumentedTest.kt
@@ -1,7 +1,7 @@
-package br.com.lucas.santanderbootcamp.todolist
+package br.com.lucas.todo
 
 import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.runner.AndroidJUnit4
 
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("br.com.lucas.santanderbootcamp.todolist", appContext.packageName)
+        assertEquals("br.com.lucas.todo", appContext.packageName)
     }
 }
