+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="br.com.lucas.santanderbootcamp.todolist">
+    package="br.com.lucas.todo">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ToDoList">
         <activity
-            android:name=".ui.customSplashScreen.CustomSplashScreenActivity"
+            android:name="br.com.lucas.todo.ui.customSplashScreen.CustomSplashScreenActivity"
             android:theme="@style/Theme.SplashScreen"
             android:exported="true">
             <intent-filter>
@@ -19,19 +19,19 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.infoTask.InfoTaskActivity"
+            android:name="br.com.lucas.todo.ui.infoTask.InfoTaskActivity"
             android:exported="true"
-            android:parentActivityName=".ui.listTask.ListTaskActivity"
+            android:parentActivityName="br.com.lucas.todo.ui.listTask.ListTaskActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".ui.listTask.ListTaskActivity"
+            android:name="br.com.lucas.todo.ui.listTask.ListTaskActivity"
             android:exported="true"
             android:screenOrientation="portrait">
         </activity>
         <activity
-            android:name=".ui.editTask.EditTaskActivity"
+            android:name="br.com.lucas.todo.ui.editTask.EditTaskActivity"
             android:exported="true"
-            android:parentActivityName=".ui.listTask.ListTaskActivity"
+            android:parentActivityName="br.com.lucas.todo.ui.listTask.ListTaskActivity"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="stateHidden" />
     </application>
