+++ app/src/main/java/br/com/lucas/todo/core/extensions/ColorExtension.kt
@@ -1,4 +1,4 @@
-package br.com.lucas.santanderbootcamp.todolist.core.extensions
+package br.com.lucas.todo.core.extensions
 
 import android.content.Context
 import android.util.TypedValue
