+++ app/src/main/java/br/com/lucas/todo/database/DataBaseConnect.kt
@@ -1,4 +1,4 @@
-package br.com.lucas.santanderbootcamp.todolist.database
+package br.com.lucas.todo.database
 
 import android.content.Context
 import androidx.room.Room
