+++ app/src/main/res/layout/activity_list_task.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.listTask.ListTaskActivity">
+    tools:context="br.com.lucas.todo.ui.listTask.ListTaskActivity">
 
     <TextView
         android:id="@+id/tv_title"
